+++ tests/hooks/hook_fails_throws_exception.py
@@ -1,0 +1,5 @@
+from inbox.models import MessageLog
+
+
+def can_send(message_log: MessageLog):
+    print(message_log.missing_prop)